change url path
redesign patient CURD,
fix schema in db.js
change url from /patient to api/patient

+++ src/controllers/patientController.js
@@ -1,9 +1,8 @@
 // import demo model 
 const {Patient} = require('../models/db.js')
 
-
 const findAll =  async (req, res) => {
-    result = await Patient.find({},{username:true,password:true})
+    result = await Patient.find()
     res.send(result)
 }
 
@@ -14,10 +13,8 @@ const findOneById = async (req, res) => {
 
 const addOne = async (req, res) => {
     console.log("adding patient")
-    const newPatient = new Patient({
-        username: req.body.username,
-        password: req.body.password
-    })
+    const newPatient = new Patient()
+    Object.assign(newPatient,req.body)
     await newPatient.save()
         .then((result) => res.send(result))
         .catch((err) => res.send(err))
@@ -25,15 +22,27 @@ const addOne = async (req, res) => {
 
 const editOne =  async (req, res) => {
     console.log("editing patient " + req.params.username)
-    await Patient.updateOne({username:req.params.username},{isGlucose:true})
-    result = await Patient.findOne({username:req.params.username},{})
-    res.send(result)
+    patient = await Patient.findOne({username:req.params.username},{})
+    if(!patient){
+        res.status(404).send("pateient not found")
+        return
+    }
+    Object.assign(patient,req.body)
+    await patient.save()
+        .then((result) => res.send(result))
+        .catch((err) => res.send(err))
+        
 }
 
 const deleteOne = async (req, res) => {
+    result = await Patient.findOne({username:req.params.username},{})
+    if(!result){
+        res.status(404).send("patient not found")
+        return
+    }
     console.log("deleting patient " + req.params.username)
     await Patient.deleteOne({username:req.params.username},{})
-    result = await Patient.find({},{username:true,password:true})
+    result = await Patient.find()
     res.send(result)
 }
